pyproject.toml


[tool.poetry]
name = "mqtt2influx"
homepage = "https://github.com/sschleemilch/mqtt2influx"
documentation = "https://github.com/sschleemilch/mqtt2influx/blob/main/README.md"
version = "1.0.0"
description = "MQTT to InfluxDB Subscribe Aggregator"
authors = ["Sebastian Schleemilch <sebastian.schleemilch@posteo.net>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
    'License :: OSI Approved :: MIT License',
]

[tool.poetry.dependencies]
python = "^3.7"
rich = "^12.5.1"
rich-click = "^1.5.1"
influxdb-client = "^1.35.0"
paho-mqtt = "^1.6.1"
pyaml = "^21.10.1"

[tool.poetry.dev-dependencies]
black = "^22.6"
mypy = "^0.971"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["src/mqtt2influx"]
show_error_codes = true
strict = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[tool.poetry.scripts]
mqtt2influx = "mqtt2influx.__main__:main"
